requirements.txt
asgiref==3.7.2
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
click==8.1.3
cryptography==2.6.1
Django==3.2.19
django-cron==0.6.0
django-crontab==0.7.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
entrypoints==0.3
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
httplib2==0.11.3
idna==2.6
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.0
Pillow==9.5.0
pycrypto==2.6.1
pycurl==7.43.0.2
PyGObject==3.30.4
PyJWT==2.8.0
PyMySQL==1.0.2
PySimpleSOAP==1.16.2
python-apt==1.8.4
python-debian==0.1.35
python-debianbts==2.8.2
pytz==2023.3
pyxdg==0.25
reportbug==7.5.2
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
SQLAlchemy==2.0.6
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.24.1
virtualenv==16.7.7
Werkzeug==2.2.3
zipp==3.15.0
